Implemented ApiCalls from Alamofire for top anime reposes and did some MVVM restructuring.
