+++ Podfile
@@ -1,0 +1,18 @@
+# Uncomment the next line to define a global platform for your project
+ platform :ios, '13.0'
+
+target 'XouTube' do
+  # Comment the next line if you don't want to use dynamic frameworks
+  use_frameworks!
+
+  # Pods for XouTube
+
+pod 'XouDevSpec', :path => '/Users/xschoeman/Documents/git/Podfolder/XouDevSpec'
+pod “Alamofire”
+
+  target 'XouTubeTests' do
+    inherit! :search_paths
+    # Pods for testing
+  end
+
+end
