+++ XouTube/Base.lproj/LaunchScreen.storyboard
@@ -15,7 +15,7 @@
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="Image" translatesAutoresizingMaskIntoConstraints="NO" id="Pv9-5c-b4X">
+                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="XouWeebLogo" translatesAutoresizingMaskIntoConstraints="NO" id="Pv9-5c-b4X">
                                 <rect key="frame" x="10" y="359.5" width="394" height="177"/>
                                 <constraints>
                                     <constraint firstAttribute="height" constant="177" id="TMu-iF-jgW"/>
@@ -59,6 +59,6 @@
         </scene>
     </scenes>
     <resources>
-        <image name="Image" width="712" height="260"/>
+        <image name="XouWeebLogo" width="154.33332824707031" height="40"/>
     </resources>
 </document>
