+++ XouTube/Base.lproj/Main.storyboard
@@ -94,10 +94,10 @@
             </objects>
             <point key="canvasLocation" x="939" y="125"/>
         </scene>
-        <!--Main Vid View Controller-->
+        <!--Main View-->
         <scene sceneID="cnk-cy-e8A">
             <objects>
-                <viewController id="xO1-nz-Hrv" customClass="MainVidViewController" customModule="XouTube" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="xO1-nz-Hrv" customClass="MainView" customModule="XouTube" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="2In-E8-Ghu">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
